requirements.txt
ffmpeg~=4.3
numpy~=1.23.5
zlib~=1.2.13
tqdm~=4.64.1
regex~=2022.10.31
transformers~=4.25.1
matplotlib~=3.6.3
seaborn~=0.12.2
librosa~=0.9.2
scikit-learn~=1.2.0
pympi-ling~=1.70.2
streamlit~=1.17.0
scipy~=1.8.1
jsonschema~=4.17.3
setuptools~=65.6.3
textgrid~=1.5